IMPROVE support for dead lettering and removal of useless checks.
Dead lettering allow a rejected message to be send to a dead letter queue registered in config.

Checks were used to verify a configuration before setting the connection with rabbitMQ. It appears useless, because using this module implies both a deep understanding of how RabbitMQ works and reading the README.md file.

Documentation on dead lettering configuration has been added.